Move owner form inputs to their own component file
