+++ packages/webcomponents/src/components/business-forms/owner-form/owner-form-inputs.tsx
@@ -1,0 +1,101 @@
+import { Component, h, Prop } from "@stencil/core";
+import { PHONE_MASKS, SSN_MASK } from '../../../utils/form-input-masks';
+import { Owner } from "../../../api/Identity";
+import { updateAddressFormValues, updateDateOfBirthFormValues, updateFormValues } from '../utils/input-handlers';
+import { FormController } from "../../form/form";
+
+@Component({
+  tag: "owner-form-inputs"
+})
+export class BusinessOwnerFormInputs {
+  @Prop() owner: Owner;
+  @Prop() ownerDefaultValue: any;
+  @Prop() errors: any;
+  @Prop() formController: FormController;
+
+  get identificationNumberLabel() {
+    return this.owner.ssn_last4 ? 'Update SSN (optional)' : 'SSN';
+  }
+
+  inputHandler = (name: string, value: string) => {
+    updateFormValues(this.formController, { [name]: value });
+  }
+
+  onAddressFormUpdate = (values: any): void => {
+    updateAddressFormValues(this.formController, {
+      ...this.formController.values.getValue().address,
+      ...values,
+    });
+  }
+  
+  render() {
+    return (
+      <div class='row gy-3'>
+        <div class="col-12 col-md-6">
+          <form-control-text
+            name="name"
+            label="Full Name"
+            defaultValue={this.ownerDefaultValue.name}
+            errorText={this.errors.name}
+            inputHandler={this.inputHandler}
+          />
+        </div>
+        <div class="col-12 col-md-4">
+          <form-control-text
+            name="title"
+            label="Title"
+            defaultValue={this.ownerDefaultValue.title}
+            errorText={this.errors.title}
+            inputHandler={this.inputHandler}
+          />
+        </div>
+        <div class="col-12 col-md-6">
+          <form-control-text
+            name="email"
+            label="Email Address"
+            defaultValue={this.ownerDefaultValue.email}
+            errorText={this.errors.email}
+            inputHandler={this.inputHandler}
+          />
+        </div>
+        <div class="col-12 col-md-6">
+          <form-control-number-masked
+            name="phone"
+            label="Phone Number"
+            defaultValue={this.ownerDefaultValue.phone}
+            errorText={this.errors.phone}
+            inputHandler={this.inputHandler}
+            mask={PHONE_MASKS.US}
+          />
+        </div>
+        <div class="col-12 col-md-4">
+          <form-control-date
+            name="dob_full"
+            label="Birth Date"
+            defaultValue={this.ownerDefaultValue.dob_full}
+            errorText={this.errors.dob_full}
+            inputHandler={this.inputHandler}
+            onFormControlInput={(e) => updateDateOfBirthFormValues(e, this.formController)}
+          />
+        </div>
+        <div class="col-12 col-md-8">
+          <form-control-number-masked
+            name="identification_number"
+            label={this.identificationNumberLabel}
+            defaultValue={this.ownerDefaultValue.identification_number}
+            errorText={this.errors.identification_number}
+            inputHandler={this.inputHandler}
+            mask={SSN_MASK}
+          />
+        </div>
+        <div class="col-12">
+          <justifi-identity-address-form
+            errors={this.errors.address}
+            defaultValues={this.ownerDefaultValue.address}
+            handleFormUpdate={this.onAddressFormUpdate}
+          />
+        </div>
+      </div>
+    )
+  };
+}
